4 separate parts to this submission. Each listed below. The new build (#2) for the Tiva chip uses files from the TivaWare_C_Series-1.0 code distribuition. The philosophy is not to distribute files that already exist elsewhere. However, this has the problem that doing a pull from this repository will not get all the files you need to do a build. If you have trouble accessing the additional files please ask and I'll either send them to you and point you to their location.
Tasks still to be done: a) Bring up the current motor code including the latest version of FreeRTOS in the new Tiva IAR build and get the motor running again with the new board. b) Get the motor control program and symbol table transferred to the controller using USB.

Submission sections:

1. Modified acceleration to make motor turn faster. This was done for the old LM3S3748 Stellaris chip. Started adding changes to _RPMx10ktoFreq(), which is not working correrectly, to use integer, rather than floating point arithmetic. Changing RPM to frequency, still needs work.

2. Added a new IAR project and workspace file for the new TI Tiva TM4C1233D5PM chip that was selected as the successor to the old Stellaris chip. The project is based on the virtual com port example provided in ti/TivaWare_C_Series-1.0/examples/boards/ek-tm4c123gxl/usb_dev_serial. Note that these examples are based on the Tiva C Series LaunchPad demo board which uses the TM4C123GH6PM chip. Chip specific assignments are dealt with in the driverlib/pin_map.h file. The #define PART_TM4C123GH6PM has been changed to the corresponding PART_TM4C1233D5PM to reflect the part used on the Sep2013 version of the motor controller board.

Note that the new IAR build is called 'motorTiva'. The old one for the Stellaris part is just called 'motor'.

3. The attempted Linux build using the gcc compiler noted in a previous submission. Were not able to get the link to succeed although the compilations went fine. There is a Yagarto sample someone has posted on the TI web site. See: http://e2e.ti.com/support/microcontrollers/tiva_arm/f/908/p/231942/815213.aspx#815213
Ironically, this addresses the USB/UART issue that came up in the development of the new board. Both the UART and USB ports are utilized in the usb_dev_serial example. The Yagarto shows how to free up using the UART port.

Note that some of the files added here (hw_nvic.h, hw_types.h, libdriver.a) are in the TivaWare_C_Series-1.0 distribution.

4. Added the first pass at a flash IO routine that is intended to be used to save the program and symbol table. This has not been tested yet.