The monster slowly awakens. Timing is off in the stepper timer isr() so the motor is barely moving. Probably due to the increased clock speed of the Tiva chip versus Stellaris. Checking in this code so we can at least get back to it if necessary. The Allegro motor driver chip seems to be working fine which is one less problem.
For debugging, try setting breakpoints at localAssert() and vApplicationStackOverflowHook(). The first will be called if any of the internal FreeRTOS configASSERT()'s trigger. Note that the #define for configASSERT() in FreeRTOSConfig.h is different from the one shown on their website: #define configASSERT( ( x ) )     if( ( x ) == 0 ) vAssertCalled( __FILE__, __LINE__ )
IAR does not find the definition amusing and neither will you when trying to debug its crypic compile time errors.

Changes here have been compiled for the Stellaris build but have not been tested on that chip.