After posting a query up on TI's e2e forum, figured out that Device Firmware Update class wasn't enumerating because the windows_drivers weren't installed. Spent a lot of time figuring this one out. There are 2 places where code is currently disabled, but when turned on will put the device in the DFU mode. See io.cpp and usb_dev_serial.c. They're each slightly different and have tested the one in io.cpp. Still need to figure out where is the best place to put this.
Turned on an xml call graph in the linker output. Thought this might be useful.

Modified the idle task to have to same priority as the interpret task. Noticed that idle was not getting time slices if interpret never yielded. Looks like the FreeRTOS model is not to bump up the priority to give it some time slices.

Started on the user's manual.