2 problems saving and retrieving program from flash memory. First issue was the location of the symbol table area. See flashio.h for the change. Note that each program entry takes 12-bytes and we have room for  3072 (0xc00) bytes so that leaves room for 256 program entries. I think a symbol table entry takes 14 bytes (but check this) and we have 1024 bytes, or 74 entries.
The other, more difficult, problem was on line 37 of flashio.cpp. We weren't saving the right amount of data from ram to flash in the first place so, of course, it was never getting restored correctly. The problem was a pointer to an int (4 bytes) rather than a pointer to a byte, which increases by 1, not 4.